+++ test/codegen/sources/aliased/schemas/Player.ts
@@ -1,0 +1,5 @@
+import { Schema, type } from "../../../../../src";
+
+export class AliasedPlayer extends Schema {
+    @type("string") name: string;
+}